Remove unused game file management code
The removed code included management of a list of currently loaded game file paths and functionality for refreshing game buttons according to this list. This code was previously commented out and has now been completely removed as it was identified as no longer necessary in the current design of the application.