Add custom URL handling and error notification
This update adds functionality for handling custom URL values for the Video Button and Info Button of the application. This includes the addition of error notifications for unsuccessful URL launches. An "Edit Links" window is now available where users can input custom URLs, with provisions for default values in case of errors or blank entries.