Update the default game image and adjust UI components
The default game image mechanism was updated to check for a default.png in the SystemImageFolder if specified; otherwise, the global default is used. Main window minimum height was increased to 500, and formatting of the interface text was improved. A label was added to display the total number of files and updated the default thumbnail size from 350 to 200. It also refactored DisplaySystemInfo method to display complete system info. The names of the emulators are displayed with their specific details. Minor modifications were made to some strings for better presentation.