Add search feature and improve interface
Implemented a feature where search can be triggered by pressing 'Enter' in MainWindow. This includes the display of warning messages when a system is not selected or a search query is missing. Also, the display messages have been updated to cater to scenarios when there are no files or no games for a particular query. Enhanced the 'Please Wait' window to include a border and a new gif image, along with the removal of the window from the taskbar. Finally, the class inheritance hierarchy of 'PleaseWaitWindow' was simplified.