Refactor error logging and alter UI updates
The changes in this commit mainly deal with how errors are logged throughout different modules of the application, replacing async calls with a new approach that uses Tasks and waiting. It also modifies some UI updates, such as the way error messages are displayed and version details are assigned. Moreover, unnecessary white spaces are removed and some methods are refactored to synchronous.