Add easy mode for system editing and emulator downloading
The commit introduces Easy Mode for editing systems and downloading emulators. In the Easy Mode, a new dropdown menu for systems and emulators is displayed. The user can choose whether they already have an emulator for the selected system. If not, a list of suitable emulators is displayed for the user to choose from and a download process can be initiated. Changes are mainly in MainWindow.xaml, EditSystemEasyMode.xaml and corresponding .cs files.