Update system configurations in launcher XML files
The changes made in this commit include updating system configurations in both 'easymodepreset.xml' and 'system_model.xml' files. This includes the addition of new systems and emulators, as well as alterations to existing system configurations. The code has been cleaned up for better management and easier understanding of system specifications.