Implement ignore list in update script and adjust resource files
The update script now has a list of files it will ignore during updates. This is primarily for files directly related to the update program itself. Also, the resource references in the SimpleLauncher project file were adjusted to not compile certain resources, these changes include adjusting the assembly version and file version.