Add favorites functionality to game launcher
Added the ability to add games to a favorites list, accessible from multiple areas of the launcher. The user can now right-click a game in the main interface or search results and add it to their favorites. The favorites page includes options to manage the list and perform actions on the favorite games. Also, updated the way paths are handled throughout the solution.