Refine search term for video and info URL
The search term for the Video and Info URL now matches a machine's description if available. Previously, the search term was just the file name without extension and the system name. Now, the machine's description is used as the search term if it matches the file name without extension, if available and not empty.