Update favorites functionality and icon in SimpleLauncher
The code changes include additional functionality to treat favourite games in a special way by displaying a star icon on them. To facilitate this, the GameButtonFactory constructor and its invocation were updated to include a FavoritesConfig argument. The Favorites icon in the MainWindow has been switched from a heart to a star. Additionally, an shutdown behaviour was added to the Favorites.xaml which, when closing, restarts the application.