Refactor game button creation, replace heart with star icon
The commit disables the use of search terms for creating video and info icons in the SimpleLauncher. It also refactors the GameButtonFactory by commenting out the methods which use the previously mentioned search term. This includes creating the VideoIcon and InfoIcon, and their respective event handlers. Finally, the heart image used in the favorites feature, both in the GameButtonFactory and MainWindow, is replaced with a star.