Update interface elements and disable message boxes
The comments in the code have disabled the success message box pop-ups in the Favorites.xaml.cs file that notify users when an item has been removed from favorites. In addition, the heart icon used in the GameButtonFactory.cs file was replaced with a star icon, projecting a change in the user interface design.