Update favorite list features and remove Video and Info link
Updated the favorite list UI feature by introducing a star icon for favourite games. Implemented a new preview tooltip feature that appears when hovering over an item in the Favorite or Global Search window. Improved the logic for creating favorites.xml and settings.xml files when they're not found. Also, the Video and Info link is now removed from the generated buttons but remains accessible through the context menu.