Update SimpleLauncher to v3.0.1 with improvements
This commit includes an update to 'SimpleLauncher', upgrading it to version 3.0.1. The changes entail bug fixes, the implementation of a file check before launching a favorite game, limiting the number of games per page to 500, and adding a second log window during the update process. These changes aim to enhance user experience and optimize performance.