Updated emulator download links and versions
This commit updates the download links and versions of the BigPEmu, DuckStation, and PCSX2 emulators within the easymode.xml file. Additionally, the updates made to the SimpleLauncher application are listed in the whatsnew.txt file, including implementation of a file check, fixing freezing issue, cap on games per page, another log window in the update process, enhanced global search functionality, and some fixes in random codes. The company and product details have also been added to the SimpleLauncher.csproj file.