Update emulator versions and download links
This commit updates the versions and download links for several emulators in the SimpleLauncher project. Changes include upgrading MAME to version 0.272, BigPEmu to 1.17, Dolphin to 2412, and BizHawk to 2.10-rc2. These updates ensure compatibility with the latest emulator releases and improve user experience by pointing to the most current download packages.